Move Journaling-related Models to App\Models\Settings\ChartOfAccounts
