+++ app/Http/Controllers/JournalVouchersController.php
@@ -3,9 +3,6 @@
 namespace App\Http\Controllers;
 
 use App\Models\JournalVouchers;
-use App\Models\JournalEntries;
-use App\Models\JournalPostings;
-use App\Models\Settings\ChartOfAccounts\ChartOfAccounts;
 use App\Actions\CreateJournalEntry;
 use App\Actions\CreateJournalPostings;
 use App\Actions\CreateJournalVoucher;
