+++ app/Models/ReceiptReferences.php
@@ -4,6 +4,7 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use App\Models\Settings\ChartOfAccounts\JournalEntries;
 
 class ReceiptReferences extends Model
 {
